[Shweta]Add: Given the States Census CSV file, Check to ensure the Number of Record matches

+++ IndianStatesCensusAnalyserProgram/DAO/CensusDataDAO.cs
@@ -15,7 +15,7 @@ namespace IndianStateCensusAnalyser
         public CensusDataDAO(string state, string population, string area, string density)
         {
             this.state = state;
-            this.population = Convert.ToInt64(population);
+            this.population = Convert.ToUInt32(population);
             this.area = Convert.ToInt64(area);
             this.density = Convert.ToInt64(density);
         }
